+++ reader/tests/test_views.py
@@ -10,13 +10,14 @@ from . import ReaderTestBase
 
 
 class ReaderViewTestBase(ReaderTestBase):
-    @staticmethod
-    def setup_series(setup_chapter: bool = True):
-        series = Series.objects.create(title='series', cover=get_test_image())
-        if setup_chapter:
-            series.chapters.create(
-                title='chapter', file=get_valid_zip_file(), number=1
-            )
+    def setup_method(self):
+        super().setup_method()
+        self.series = Series.objects.create(
+            title='series', cover=get_test_image()
+        )
+        self.series.chapters.create(
+            title='chapter', file=get_valid_zip_file(), number=1
+        )
 
     def teardown_method(self):
         super().teardown_method()
@@ -27,26 +28,23 @@ class TestDirectory(ReaderViewTestBase):
     URL = reverse('reader:directory')
 
     def test_get(self):
-        self.setup_series()
         r = self.client.get(self.URL)
         assert r.status_code == 200
 
 
 class TestSeries(ReaderViewTestBase):
     def test_get(self):
-        self.setup_series()
         url = reverse('reader:series', kwargs={'slug': 'series'})
         r = self.client.get(url)
         assert r.status_code == 200
 
     def test_get_not_found(self):
-        self.setup_series()
         url = reverse('reader:series', kwargs={'slug': 'not-found'})
         r = self.client.get(url)
         assert r.status_code == 404
 
     def test_get_unavailable(self):
-        self.setup_series(setup_chapter=False)
+        self.series.chapters.all().delete()
         self.client.logout()
         url = reverse('reader:series', kwargs={'slug': 'series'})
         r = self.client.get(url)
@@ -55,7 +53,6 @@ class TestSeries(ReaderViewTestBase):
 
 class TestChapterPage(ReaderViewTestBase):
     def test_get(self):
-        self.setup_series()
         url = reverse('reader:page', kwargs={
             'slug': 'series', 'vol': 0, 'num': 1, 'page': 1
         })
@@ -63,7 +60,6 @@ class TestChapterPage(ReaderViewTestBase):
         assert r.status_code == 200
 
     def test_get_page_zero(self):
-        self.setup_series()
         url = reverse('reader:page', kwargs={
             'slug': 'series', 'vol': 0, 'num': 1, 'page': 0
         })
@@ -71,7 +67,6 @@ class TestChapterPage(ReaderViewTestBase):
         assert r.status_code == 404
 
     def test_get_page_not_found(self):
-        self.setup_series()
         url = reverse('reader:page', kwargs={
             'slug': 'series', 'vol': 1, 'num': 1, 'page': 3
         })
@@ -81,7 +76,6 @@ class TestChapterPage(ReaderViewTestBase):
 
 class TestChapterRedirect(ReaderViewTestBase):
     def test_get(self):
-        self.setup_series()
         url = reverse('reader:chapter', kwargs={
             'slug': 'series', 'vol': 0, 'num': 1
         })
@@ -99,7 +93,6 @@ class TestChapterRedirect(ReaderViewTestBase):
 
 class TestChapterDownload(ReaderViewTestBase):
     def test_get(self):
-        self.setup_series()
         url = reverse('reader:cbz', kwargs={
             'slug': 'series', 'vol': 0, 'num': 1
         })
@@ -109,7 +102,7 @@ class TestChapterDownload(ReaderViewTestBase):
         assert r.filename.endswith('c1.cbz')
 
     def test_get_not_found(self):
-        self.setup_series(setup_chapter=False)
+        self.series.chapters.all().delete()
         url = reverse('reader:cbz', kwargs={
             'slug': 'series', 'vol': 0, 'num': 1
         })